# Front End with code/Front End - React/ai-sdr-frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI SDR Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/meetings/requests": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Create Request",
        "operationId": "create_request_meetings_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingBooking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Meetings",
        "operationId": "list_meetings_meetings_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Get Meeting",
        "operationId": "get_meeting_meetings__meeting_id__get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingBooking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}/approve-and-book": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Approve And Book",
        "operationId": "approve_and_book_meetings__meeting_id__approve_and_book_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingBooking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}/cancel": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Cancel Meeting",
        "operationId": "cancel_meeting_meetings__meeting_id__cancel_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingBooking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}/reminders/{reminder_type}": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Send Reminder",
        "operationId": "send_reminder_meetings__meeting_id__reminders__reminder_type__post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          },
          {
            "name": "reminder_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reminder Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingBooking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}/retry-crm": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Retry Crm",
        "operationId": "retry_crm_meetings__meeting_id__retry_crm_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingBooking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/webhooks/provider": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Provider Webhook",
        "operationId": "provider_webhook_meetings_webhooks_provider_post",
        "parameters": [
          {
            "name": "x-sdr-webhook-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sdr-Webhook-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingBooking"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeetingBooking": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "title": "Meeting Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id"
          },
          "intelligence_id": {
            "type": "string",
            "title": "Intelligence Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "contact_id": {
            "type": "string",
            "title": "Contact Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "contact_name": {
            "type": "string",
            "title": "Contact Name"
          },
          "attendee_email": {
            "type": "string",
            "title": "Attendee Email"
          },
          "organizer_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organizer Email"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "duration_minutes": {
            "type": "integer",
            "maximum": 240.0,
            "minimum": 15.0,
            "title": "Duration Minutes",
            "default": 30
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          },
          "status": {
            "type": "string",
            "enum": [
              "awaiting_approval",
              "approved",
              "booking",
              "booked",
              "cancelled",
              "failed"
            ],
            "title": "Status",
            "default": "awaiting_approval"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "dry_run"
          },
          "provider_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Event Id"
          },
          "meeting_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Url"
          },
          "calendar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Url"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "crm_sync_status": {
            "type": "string",
            "enum": [
              "not_started",
              "synced",
              "failed"
            ],
            "title": "Crm Sync Status",
            "default": "not_started"
          },
          "crm_sync_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crm Sync Id"
          },
          "temporal_workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporal Workflow Id"
          },
          "temporal_status": {
            "type": "string",
            "enum": [
              "not_started",
              "started",
              "failed"
            ],
            "title": "Temporal Status",
            "default": "not_started"
          },
          "provider_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provider Payload"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "campaign_id",
          "intelligence_id",
          "account_id",
          "contact_id",
          "company_name",
          "contact_name",
          "attendee_email",
          "title"
        ],
        "title": "MeetingBooking"
      },
      "MeetingCancelRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": "Cancelled by SDR"
          }
        },
        "type": "object",
        "title": "MeetingCancelRequest"
      },
      "MeetingDraftRequest": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          }
        },
        "type": "object",
        "required": [
          "conversation_id"
        ],
        "title": "MeetingDraftRequest"
      },
      "MeetingListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeetingBooking"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "MeetingListResponse"
      },
...
      "MeetingScheduleRequest": {
        "properties": {
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "duration_minutes": {
            "type": "integer",
            "maximum": 240.0,
            "minimum": 15.0,
            "title": "Duration Minutes",
            "default": 30
          },
          "approved_by": {
            "type": "string",
            "title": "Approved By"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "start_at",
          "approved_by"
        ],
        "title": "MeetingScheduleRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}